Rename segway_input_master
